requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure OpenAI
openai>=1.0.0

# Application Insights & OpenTelemetry
azure-monitor-opentelemetry>=1.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
